@@ -1,15 +1,13 @@
 [build-system]
-# change this to be "setuptools>=77.0.3" after 3.8 is out of support
-requires = ["setuptools>=61.2"]
+requires = ["setuptools>=77.0.3"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyfakefs"
 description = "Implements a fake file system that mocks the Python file system modules."
 readme = "README.md"
-# change this to be PEP-639-conform after 3.8 is out of support
-license = {file = "COPYING"}
-
+license = "Apache-2.0"
+license-files = ["COPYING"]
 authors = [
     { name = "Google", email = "google-pyfakefs@google.com" },
     { name = "John McGehee", email = "pyfakefs@johnnado.com" }
@@ -21,11 +19,7 @@ classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Environment :: Console",
     "Intended Audience :: Developers",
-    "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -49,7 +43,7 @@ keywords = [
     "unittest",
     "pytest",
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.10"
 dynamic = ["version"]
 entry-points.pytest11.fakefs = "pyfakefs.pytest_plugin"
 
@@ -69,12 +63,9 @@ doc = [
     "furo>=2024.8.6"
 ]
 extra = [
-    "pandas==1.3.5; python_version == '3.7'",
-    "pandas==2.0.3; python_version == '3.8'",
-    "pandas==2.2.3; python_version > '3.8'",
+    "pandas==2.2.3",
     "xlrd==2.0.1",
-    "openpyxl==3.1.3; python_version == '3.7'",
-    "openpyxl==3.1.5; python_version > '3.7'",
+    "openpyxl==3.1.5",
 ]
 xdist = [
     "pytest-xdist",
